Refactored the search input
